Move filters to read.go
