+++ provider/aclrule/datasource_rediscloud_acl_rule.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/RedisLabs/rediscloud-go-api/service/access_control_lists/redis_rules"
 
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
@@ -70,21 +69,3 @@ func (d *aclRuleDataSource) Configure(_ context.Context, req datasource.Configur
 	d.client = client
 }
 
-func filterRules(list []*redis_rules.GetRedisRuleResponse, filters []func(*redis_rules.GetRedisRuleResponse) bool) []*redis_rules.GetRedisRuleResponse {
-	var filtered []*redis_rules.GetRedisRuleResponse
-	for _, rule := range list {
-		if filterRule(rule, filters) {
-			filtered = append(filtered, rule)
-		}
-	}
-	return filtered
-}
-
-func filterRule(rule *redis_rules.GetRedisRuleResponse, filters []func(*redis_rules.GetRedisRuleResponse) bool) bool {
-	for _, filter := range filters {
-		if !filter(rule) {
-			return false
-		}
-	}
-	return true
-}
